Minor format changes and addition of comments

+++ Resources/script.js
@@ -8,7 +8,7 @@ const total_btn = document.querySelector('#total-btn')
 tip_btn.addEventListener('click', calc_tip);
 
 function calc_tip () {
- let bill = parseInt(bill_input.value);
+ let bill = parseFloat(bill_input.value);
  let percentage = parseInt(percentage_input.value) / 100;
  let people = parseInt(people_input.value);
  let tip = (bill * percentage) / people;
@@ -16,6 +16,7 @@ function calc_tip () {
  
  let tip_calc = document.querySelector('.tip-section .tip');
  tip_calc.innerText = '$' + tip.toFixed(2) + ((people > 1) ? ' per person' : '');
+ //Different format for loop found on Stacks Overflow
 
  let total = document.getElementById('totality');
  total.innerHTML = '$' + billtotal.toFixed(2)
